:hammer: Dedupe containsVersionMismatch function
