+++ ern-core/test/resolveNativeDependenciesVersions-test.ts
@@ -1,7 +1,69 @@
 import { expect } from 'chai'
 import { NativeDependencies } from '../src/nativeDependenciesLookup'
 import { PackagePath } from '../src/PackagePath'
-import * as nativeDepenciesVersionResolution from '../src/resolveNativeDependenciesVersions'
+import {
+  containsVersionMismatch,
+  resolveNativeDependenciesVersions,
+  retainHighestVersions,
+} from '../src/resolveNativeDependenciesVersions'
+
+// ==========================================================
+// containsVersionMismatch
+// ==========================================================
+const versionsWithAMajorMismatch = ['1.0.0', '2.0.0', '1.0.0']
+const versionsWithAMinorMismatch = ['1.0.0', '1.1.0', '1.0.0']
+const versionsWithAPatchMismatch = ['1.0.0', '1.0.1', '1.0.0']
+const versionsWithoutMismatch = ['1.0.0', '1.0.0', '1.0.0']
+
+describe('containsVersionMismatch', () => {
+  it('should return true if mismatch level is set to major and there is at least one major version mismatch', () => {
+    expect(containsVersionMismatch(versionsWithAMajorMismatch, 'major')).true
+  })
+
+  it('should return false if mismatch level is set to major and there is no major version mismatch [1]', () => {
+    expect(containsVersionMismatch(versionsWithAMinorMismatch, 'major')).false
+  })
+
+  it('should return false if mismatch level is set to major and there is no major version mismatch [2]', () => {
+    expect(containsVersionMismatch(versionsWithAPatchMismatch, 'major')).false
+  })
+
+  it('should return false if mismatch level is set to major and there is no major version mismatch [3]', () => {
+    expect(containsVersionMismatch(versionsWithoutMismatch, 'major')).false
+  })
+
+  it('should return true if mismatch level is set to minor and there is at least one major version mismatch', () => {
+    expect(containsVersionMismatch(versionsWithAMajorMismatch, 'minor')).true
+  })
+
+  it('should return true if mismatch level is set to minor and there is at least one minor version mismatch', () => {
+    expect(containsVersionMismatch(versionsWithAMinorMismatch, 'minor')).true
+  })
+
+  it('should return false if mismatch level is set to minor and there no minor version mismatch [1]', () => {
+    expect(containsVersionMismatch(versionsWithAPatchMismatch, 'minor')).false
+  })
+
+  it('should return false if mismatch level is set to minor and there no minor version mismatch [1]', () => {
+    expect(containsVersionMismatch(versionsWithoutMismatch, 'minor')).false
+  })
+
+  it('should return true if mismatch level is set to patch and there is at least one major version mismatch', () => {
+    expect(containsVersionMismatch(versionsWithAMajorMismatch, 'patch')).true
+  })
+
+  it('should return true if mismatch level is set to patch and there is at least one minor version mismatch', () => {
+    expect(containsVersionMismatch(versionsWithAMinorMismatch, 'patch')).true
+  })
+
+  it('should return true if mismatch level is set to patch and there is at least one patch version mismatch', () => {
+    expect(containsVersionMismatch(versionsWithAPatchMismatch, 'patch')).true
+  })
+
+  it('should return false if mismatch level is set to patch and there is no patch version mismatch', () => {
+    expect(containsVersionMismatch(versionsWithoutMismatch, 'patch')).false
+  })
+})
 
 // ==========================================================
 // retainHighestVersions
@@ -23,10 +85,7 @@ describe('retainHighestVersion', () => {
       PackagePath.fromString('dependencyF@3.0.0'),
     ]
 
-    const result = nativeDepenciesVersionResolution.retainHighestVersions(
-      arrA,
-      arrB
-    )
+    const result = retainHighestVersions(arrA, arrB)
     const stringifedResult = result.map(p => p.toString())
     expect(stringifedResult).length(6)
     expect(stringifedResult).includes('dependencyA@2.0.0')
@@ -89,9 +148,7 @@ describe('resolveNativeDependenciesVersions', () => {
       },
     ]
 
-    const result = nativeDepenciesVersionResolution.resolveNativeDependenciesVersions(
-      fixture
-    )
+    const result = resolveNativeDependenciesVersions(fixture)
     expect(result.resolved).length(4)
     expect(result.pluginsWithMismatchingVersions).empty
   })
@@ -143,9 +200,7 @@ describe('resolveNativeDependenciesVersions', () => {
       },
     ]
 
-    const result = nativeDepenciesVersionResolution.resolveNativeDependenciesVersions(
-      fixture
-    )
+    const result = resolveNativeDependenciesVersions(fixture)
     expect(result.resolved).length(3)
     expect(result.pluginsWithMismatchingVersions).empty
     const resolvedDepsAsStrings = result.resolved.map(r => r.toString())
@@ -199,9 +254,7 @@ describe('resolveNativeDependenciesVersions', () => {
       },
     ]
 
-    const result = nativeDepenciesVersionResolution.resolveNativeDependenciesVersions(
-      fixture
-    )
+    const result = resolveNativeDependenciesVersions(fixture)
     expect(result.resolved).length(3)
     expect(result.pluginsWithMismatchingVersions).empty
     const resolvedDepsAsStrings = result.resolved.map(r => r.toString())
@@ -255,9 +308,7 @@ describe('resolveNativeDependenciesVersions', () => {
       },
     ]
 
-    const result = nativeDepenciesVersionResolution.resolveNativeDependenciesVersions(
-      fixture
-    )
+    const result = resolveNativeDependenciesVersions(fixture)
     expect(result.resolved).length(3)
   })
 
@@ -308,9 +359,7 @@ describe('resolveNativeDependenciesVersions', () => {
       },
     ]
 
-    const result = nativeDepenciesVersionResolution.resolveNativeDependenciesVersions(
-      fixture
-    )
+    const result = resolveNativeDependenciesVersions(fixture)
     expect(result.resolved).length(2)
     expect(result.pluginsWithMismatchingVersions).length(1)
     expect(result.pluginsWithMismatchingVersions).includes('apiOne')
@@ -363,9 +412,7 @@ describe('resolveNativeDependenciesVersions', () => {
       },
     ]
 
-    const result = nativeDepenciesVersionResolution.resolveNativeDependenciesVersions(
-      fixture
-    )
+    const result = resolveNativeDependenciesVersions(fixture)
     expect(result.resolved).length(3)
     expect(result.pluginsWithMismatchingVersions).length(0)
     const resolvedDepsAsStrings = result.resolved.map(r => r.toString())
@@ -419,9 +466,7 @@ describe('resolveNativeDependenciesVersions', () => {
       },
     ]
 
-    const result = nativeDepenciesVersionResolution.resolveNativeDependenciesVersions(
-      fixture
-    )
+    const result = resolveNativeDependenciesVersions(fixture)
     expect(result.resolved).length(2)
     expect(result.pluginsWithMismatchingVersions).length(1)
     expect(result.pluginsWithMismatchingVersions).includes('nativeModuleOne')
@@ -476,9 +521,7 @@ describe('resolveNativeDependenciesVersions', () => {
       },
     ]
 
-    const result = nativeDepenciesVersionResolution.resolveNativeDependenciesVersions(
-      fixture
-    )
+    const result = resolveNativeDependenciesVersions(fixture)
     expect(result.resolved).length(2)
     expect(result.pluginsWithMismatchingVersions).length(1)
     expect(result.pluginsWithMismatchingVersions).includes('nativeModuleOne')
@@ -533,9 +576,7 @@ describe('resolveNativeDependenciesVersions', () => {
       },
     ]
 
-    const result = nativeDepenciesVersionResolution.resolveNativeDependenciesVersions(
-      fixture
-    )
+    const result = resolveNativeDependenciesVersions(fixture)
     expect(result.resolved).length(2)
     expect(result.pluginsWithMismatchingVersions).length(1)
     expect(result.pluginsWithMismatchingVersions).includes('nativeModuleOne')
