--- ern-orchestrator/test/publication-test.js
@@ -1,91 +1,0 @@
-import mockFs from 'mock-fs'
-import os from 'os'
-import path from 'path'
-import { assert, expect } from 'chai'
-import { PackagePath } from 'ern-core'
-import { getCodePushSdk, getCodePushInitConfig } from '../src/codepush'
-import {
-  containsVersionMismatch,
-  resolvePluginsVersions,
-} from '../src/container'
-
-const ernRcPath = path.join(os.homedir(), '.ern', '.ernrc')
-const codePushConfigPath = path.join(
-  process.env.LOCALAPPDATA || process.env.HOME || '',
-  '.code-push.config'
-)
-const ernRcCodePushAccessKey = '0e2509c78c4f94c25e69131a0a5e5be3b7d2927b'
-const codePushConfigAccessKey = '1e2509c78c4f94c25e69131a0a5e5be3b7d2927b'
-
-const ernRcWithCodePushAccessKey = JSON.stringify({
-  codePushAccessKey: ernRcCodePushAccessKey,
-})
-const codePushConfigWithAccessKey = JSON.stringify({
-  accessKey: codePushConfigAccessKey,
-})
-const ernRcWithoutCodePushAccessKey = JSON.stringify({})
-const codePushConfigWithoutAccessKey = JSON.stringify({})
-
-describe('publication.js', () => {
-  afterEach(() => {
-    mockFs.restore()
-  })
-
-  // ==========================================================
-  // containsVersionMismatch
-  // ==========================================================
-  const versionsWithAMajorMismatch = ['1.0.0', '2.0.0', '1.0.0']
-  const versionsWithAMinorMismatch = ['1.0.0', '1.1.0', '1.0.0']
-  const versionsWithAPatchMismatch = ['1.0.0', '1.0.1', '1.0.0']
-  const versionsWithoutMismatch = ['1.0.0', '1.0.0', '1.0.0']
-
-  describe('containsVersionMismatch', () => {
-    it('should return true if mismatch level is set to major and there is at least one major version mismatch', () => {
-      expect(containsVersionMismatch(versionsWithAMajorMismatch, 'major')).true
-    })
-
-    it('should return false if mismatch level is set to major and there is no major version mismatch [1]', () => {
-      expect(containsVersionMismatch(versionsWithAMinorMismatch, 'major')).false
-    })
-
-    it('should return false if mismatch level is set to major and there is no major version mismatch [2]', () => {
-      expect(containsVersionMismatch(versionsWithAPatchMismatch, 'major')).false
-    })
-
-    it('should return false if mismatch level is set to major and there is no major version mismatch [3]', () => {
-      expect(containsVersionMismatch(versionsWithoutMismatch, 'major')).false
-    })
-
-    it('should return true if mismatch level is set to minor and there is at least one major version mismatch', () => {
-      expect(containsVersionMismatch(versionsWithAMajorMismatch, 'minor')).true
-    })
-
-    it('should return true if mismatch level is set to minor and there is at least one minor version mismatch', () => {
-      expect(containsVersionMismatch(versionsWithAMinorMismatch, 'minor')).true
-    })
-
-    it('should return false if mismatch level is set to minor and there no minor version mismatch [1]', () => {
-      expect(containsVersionMismatch(versionsWithAPatchMismatch, 'minor')).false
-    })
-
-    it('should return false if mismatch level is set to minor and there no minor version mismatch [1]', () => {
-      expect(containsVersionMismatch(versionsWithoutMismatch, 'minor')).false
-    })
-
-    it('should return true if mismatch level is set to patch and there is at least one major version mismatch', () => {
-      expect(containsVersionMismatch(versionsWithAMajorMismatch, 'patch')).true
-    })
-
-    it('should return true if mismatch level is set to patch and there is at least one minor version mismatch', () => {
-      expect(containsVersionMismatch(versionsWithAMinorMismatch, 'patch')).true
-    })
-
-    it('should return true if mismatch level is set to patch and there is at least one patch version mismatch', () => {
-      expect(containsVersionMismatch(versionsWithAPatchMismatch, 'patch')).true
-    })
-
-    it('should return false if mismatch level is set to patch and there is no patch version mismatch', () => {
-      expect(containsVersionMismatch(versionsWithoutMismatch, 'patch')).false
-    })
-  })
-})
